// console/src/generated/UsageStatisticsDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type UsageStatisticsDto = { user_key_id: number | null, user_id: number | null, provider_id: number | null, model: string | null, requests: number, input_tokens: number, output_tokens: number, cached_input_tokens: number, cache_creation_5m_tokens: number, cache_creation_30m_tokens: number, cache_creation_1h_tokens: number, cost: string, };